+++ example/android/app/src/main/kotlin/com/appsflyer/example/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.appsflyer.example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}