+++ lib/src/appsflyer_sdk.dart
@@ -2,6 +2,7 @@ part of appsflyer_sdk;
 
 class AppsflyerSdk {
   StreamController _afGCDStreamController;
+  StreamController _afUDLStreamController;
   StreamController _afOpenAttributionStreamController;
   StreamController _afValidtaPurchaseController;
   EventChannel _eventChannel;
@@ -83,6 +84,11 @@ class AppsflyerSdk {
       validatedOptions[AppsflyerConstants.AF_GCD] = false;
     }
 
+    if (_afUDLStreamController != null ) {
+      validatedOptions[AppsflyerConstants.AF_UDL] = true;
+    } else {
+      validatedOptions[AppsflyerConstants.AF_UDL] = false;
+    }
     return validatedOptions;
   }
 
@@ -145,6 +151,12 @@ class AppsflyerSdk {
       afOptions[AppsflyerConstants.AF_GCD] = false;
     }
 
+    if (_afUDLStreamController != null ) {
+      afOptions[AppsflyerConstants.AF_UDL] = true;
+    } else {
+      afOptions[AppsflyerConstants.AF_UDL] = false;
+    }
+
     return afOptions;
   }
 
@@ -170,8 +182,23 @@ class AppsflyerSdk {
     }
   }
 
+
   Stream<Map> get appOpenAttributionStream {
     return _afOpenAttributionStreamController?.stream?.asBroadcastStream();
+  }
+
+  // Unified deeplink: Accessing AppsFlyer deeplink attributes
+  void _registerUDLCallback() {
+    if (_afUDLStreamController == null) {
+      _afUDLStreamController = StreamController<Map>(onCancel: () {
+        _afUDLStreamController.close();
+      });
+      _registerUDLListener();
+    }
+  }
+
+  Stream<Map> get onDeepLinkingStream {
+    return _afUDLStreamController?.stream?.asBroadcastStream();
   }
 
   ///Returns `Stream`. Accessing AppsFlyer purchase validation data
@@ -189,8 +216,10 @@ class AppsflyerSdk {
   ///initialize the SDK, using the options initialized from the constructor|
   Future<dynamic> initSdk(
       {bool registerConversionDataCallback = false,
-      bool registerOnAppOpenAttributionCallback = false}) async {
+      bool registerOnAppOpenAttributionCallback = false,
+      bool registerOnDeepLinkingCallback = false}) async {
     return Future.delayed(Duration(seconds: 0)).then((_) {
+
       if (registerConversionDataCallback) _registerConversionDataCallback();
       if (registerOnAppOpenAttributionCallback)
         _registerOnAppOpenAttributionCallback();
@@ -199,6 +228,10 @@ class AppsflyerSdk {
           registerOnAppOpenAttributionCallback) {
         _registerGCDListener();
       }
+
+      if (registerOnDeepLinkingCallback) {
+        _registerUDLCallback();
+      }      
 
       Map<String, dynamic> validatedOptions;
       if (mapOptions != null) {
@@ -288,7 +321,7 @@ class AppsflyerSdk {
 
   /// Once this API is invoked, our SDK no longer communicates with our servers and stops functioning.
   /// In some extreme cases you might want to shut down all SDK activity due to legal and privacy compliance.
-  /// This can be achieved with the stopTracking API.
+  /// This can be achieved with the stop API.
   void stop(bool isStopped) {
     _methodChannel.invokeMethod("stop", {'isStopped': isStopped});
   }
@@ -352,6 +385,25 @@ class AppsflyerSdk {
   void setAdditionalData(Map<String, dynamic> customData) {
     _methodChannel
         .invokeMethod("setAdditionalData", {'customData': customData});
+  }
+
+  void _registerUDLListener() {
+    _eventChannel.receiveBroadcastStream().listen((data) { 
+      var decodedJSON = jsonDecode(data);
+      String type = decodedJSON['type'];
+      if(type == AppsflyerConstants.AF_ON_DEEP_LINK){
+        if (_afUDLStreamController != null &&
+              !_afUDLStreamController.isClosed) {
+            _afUDLStreamController.sink.add(decodedJSON);
+        } else {
+          if ((afOptions != null && afOptions.showDebug) ||
+              (mapOptions != null &&
+                  mapOptions[AppsflyerConstants.AF_IS_DEBUG])) {
+            print("UDL Stream controller is closed. the event wasn't sent");
+          }
+        }
+      }
+    });
   }
 
   void _registerGCDListener() {
@@ -467,11 +519,19 @@ class AppsflyerSdk {
     _methodChannel.invokeMethod("setOneLinkCustomDomain", brandDomains);
   }
 
+  void setPushNotification(bool isEnabled) {
+    _methodChannel.invokeMethod("setPushNotification", isEnabled);
+  }
+
   void onInstallConversionData(Function callback) async {
     startListening(callback, "onInstallConversionData");
   }
 
   void onAppOpenAttribution(Function callback) async {
     startListening(callback, "onAppOpenAttribution");
+  }
+
+  void onDeepLinking(Function callback) async {
+    startListening(callback, "onDeepLinking");
   }
 }