+++ CHANGELOG.md
@@ -1,7 +1,25 @@
 # Versions
+
 ## 6.2.1+2
 - Hot Fix
-## 6.0.5+4
+
+## 6.2.1+1
+- Added support for push notification API
+
+## 6.2.1
+- Update iOS to v6.2.1
+- Added support for Unified Deeplink
+- Fixed deeplinks issues both for Android & iOS
+
+## 6.2.0+2
+- Revert back to version 6.2.0
+
+## 6.2.0+1
+- Added Unified Deeplinking for Android
+
+## 6.2.0
+- Update both iOS & Android to v6.2.0
+
 ## 6.0.5+3
 - Fixed `FormatException` caused by iOS side